requirements_modified.txt
 
absl-py==2.1.0
addict==2.4.0
astra-toolbox==2.1.2
beautifulsoup4==4.12.3
blobfile==2.1.1
cachetools==5.3.3
certifi==2023.11.17
charset-normalizer==3.3.2
docopt==0.6.2
einops==0.7.0
facexlib==0.3.0
filelock==3.13.1
filterpy==1.4.5
fonttools==4.25.0
fsspec==2024.3.1
ftfy==6.1.1
gdown==5.1.0
google-auth==2.29.0
google-auth-oauthlib==0.4.6
grpcio==1.62.1
h5py==3.10.0
huggingface-hub==0.22.2
idna==3.6
imageio==2.33.1
imgaug==0.4.0
importlib_metadata==7.1.0
Jinja2==3.1.3
joblib==1.4.0
lazy_loader==0.3
llvmlite==0.42.0
lmdb==1.4.1
lpips==0.1.4
lxml==4.9.4
Markdown==3.6
MarkupSafe==2.1.4
munkres==1.1.4
natsort==8.4.0
networkx==3.2.1
numba==0.59.1
numpy==1.26.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
openai-clip==1.0.1
opencv-python==4.9.0.80
pandas==2.2.1
Pillow==9.2.0
platformdirs==4.2.0
ply==3.11
protobuf==3.19.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodomex==3.20.0
pyipa==1.2.0
pyiqa==0.1.4
pyproject==1.3.1
PyQt5==5.15.10
PySocks==1.7.1
pytz==2024.1
PyYAML==6.0.1
regex==2022.8.17
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.2
scipy==1.10.1
shapely==2.0.3
soupsieve==2.5
tb-nightly==2.11.0a20220906
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.4.0
tifffile==2024.2.12
timm==0.9.16
torch==1.13.1
torchvision==0.14.1
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.1.0
wcwidth==0.2.13
Werkzeug==3.0.2
yacs==0.1.8
yapf==0.40.2

